requirements.txt
torch 
torchvision 
torchaudio
pyyaml
tqdm
Pillow
opencv-python
pycocotools
cython_bbox
lap
ultralytics